+++ CHANGES.rst
@@ -43,8 +43,9 @@ New Features
 
   - Add Planck 2015 cosmology [#3476]
 
-  - Distance calculations in cosmologies with massless (or no) neutrinos
-    are now > 10x faster for the supplied cosmologies. [#3767]
+  - Distance calculations now > 20-40x faster for the supplied
+    cosmologies due to implementing cython scalar versions of
+    ``FLRW.inv_efunc``.[#4127]
 
   - ``FLRW._tfunc`` and ``FLRW._xfunc`` are marked as deprecated.  Users
     should use the new public interfaces ``FLRW.lookback_time_integrand``